+++ src/shm_bbuffer_spmc_test.cc
@@ -48,7 +48,7 @@ int main(int argc, const char *argv[]) {
         print_usage_and_exit(app);
     const std::string app_kind = argv[1];
     const char *shm_name = argv[2];
-    const idx_t capacity = std::stod(argv[3]);
+    const idx_t capacity = std::stoul(argv[3]);
     if (capacity <= 0) {
         std::cerr << "invalid capacity\n";
         exit(EXIT_FAILURE);
